0819 write

+++ src/main/java/com/example/uscoding/repository/UserAccountRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.math.BigInteger;
 import java.util.Optional;
 
-public interface UserAccountRepository extends JpaRepository<UserAccount, BigInteger> {
+public interface UserAccountRepository extends JpaRepository<UserAccount, Long> {
     Optional<UserAccount> findByEmail(String email);
     Optional<UserAccount> findByNickname(String nickname);
 }
